[Fix] 구인 글 작성 컴포넌트 별 분리
